Inline filter implementation with active working
TODO:
  - figure out what to do with inline filter model
  - fix title for sort button and secondary action button
  - docs
  - demo
