+++ src/modules/list-data-provider-in-memory/list-data-in-memory.provider.ts
@@ -90,7 +90,7 @@ export class SkyListInMemoryDataProvider extends ListDataProvider {
         result = result.filter(item => {
           for (let i = 0; i < filters.length; i++) {
             let filter = filters[i];
-            if (filter.value === undefined || filter.value === '') {
+            if (filter.value === undefined || filter.value === '' || filter.value === false) {
               continue;
             }
