+++ src/modules/list-filters/index.ts
@@ -1,3 +1,6 @@
 export { SkyListFiltersModule } from './list-filters.module';
 export { SkyListFilterButtonComponent } from './list-filter-button.component';
 export { SkyListFilterSummaryComponent } from './list-filter-summary.component';
+export { SkyListFilterInlineComponent } from './list-filter-inline.component';
+export { SkyListFilterInlineItemComponent } from './list-filter-inline-item.component';
+
